Deprecate UserIdHasher class
